rename file handler
